[exec-if] step 2: create executor-types crate
