+++ consensus/consensus-types/src/executed_block.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{block::Block, common::Round, quorum_cert::QuorumCert};
-use executor::{ExecutedTrees, ProcessedVMOutput, StateComputeResult};
+use executor_types::{ExecutedTrees, ProcessedVMOutput, StateComputeResult};
 use libra_crypto::hash::HashValue;
 use libra_types::block_info::BlockInfo;
 use std::{
