+++ consensus/src/chained_bft/block_storage/block_store.rs
@@ -15,7 +15,7 @@ use consensus_types::{
     timeout_certificate::TimeoutCertificate,
 };
 use debug_interface::event;
-use executor::{ExecutedTrees, ProcessedVMOutput};
+use executor_types::{ExecutedTrees, ProcessedVMOutput};
 use libra_crypto::HashValue;
 use libra_logger::prelude::*;
