+++ execution/executor-types/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "executor-types"
+version = "0.1.0"
+authors = ["Libra Association <opensource@libra.org>"]
+description = "Libra executor interface types"
+repository = "https://github.com/libra/libra"
+homepage = "https://libra.org"
+license = "Apache-2.0"
+publish = false
+edition = "2018"
+
+[dependencies]
+serde = { version = "1.0.99", default-features = false }
+
+libra-crypto = { path = "../../crypto/crypto", version = "0.1.0" }
+libra-types = { path = "../../types", version = "0.1.0" }
+scratchpad = { path = "../../storage/scratchpad", version = "0.1.0" }
+storage-proto = { path = "../../storage/storage-proto", version = "0.1.0" }
+
+[features]
+default = []
+fuzzing = ["libra-crypto/fuzzing", "libra-types/fuzzing"]
