+++ execution/executor/src/lib.rs
@@ -12,13 +12,11 @@ mod mock_vm;
 pub mod utils;
 
 use anyhow::{bail, ensure, format_err, Result};
+use executor_types::{ExecutedTrees, ProcessedVMOutput, ProofReader, TransactionData};
 use futures::executor::block_on;
 use libra_config::config::{NodeConfig, VMConfig};
 use libra_crypto::{
-    hash::{
-        CryptoHash, EventAccumulatorHasher, TransactionAccumulatorHasher,
-        ACCUMULATOR_PLACEHOLDER_HASH, SPARSE_MERKLE_PLACEHOLDER_HASH,
-    },
+    hash::{CryptoHash, EventAccumulatorHasher},
     HashValue,
 };
 use libra_logger::prelude::*;
@@ -26,7 +24,6 @@ use libra_types::{
     account_address::AccountAddress,
     account_state::AccountState,
     account_state_blob::AccountStateBlob,
-    contract_event::ContractEvent,
     crypto_proxies::{LedgerInfoWithSignatures, ValidatorSet},
     proof::{accumulator::InMemoryAccumulator, definition::LeafCount, SparseMerkleProof},
     transaction::{
@@ -36,8 +33,7 @@ use libra_types::{
     write_set::{WriteOp, WriteSet},
 };
 use once_cell::sync::Lazy;
-use scratchpad::{ProofRead, SparseMerkleTree};
-use serde::{Deserialize, Serialize};
+use scratchpad::SparseMerkleTree;
 use std::{
     collections::{hash_map, HashMap, HashSet},
     convert::TryFrom,
@@ -45,234 +41,11 @@ use std::{
     sync::Arc,
 };
 use storage_client::{StorageRead, StorageWrite, VerifiedStateView};
-use storage_proto::TreeState;
 use tokio::runtime::Runtime;
 use vm_runtime::VMExecutor;
 
 static OP_COUNTERS: Lazy<libra_metrics::OpMetrics> =
     Lazy::new(|| libra_metrics::OpMetrics::new_and_registered("executor"));
-
-/// A structure that summarizes the result of the execution needed for consensus to agree on.
-/// The execution is responsible for generating the ID of the new state, which is returned in the
-/// result.
-///
-/// Not every transaction in the payload succeeds: the returned vector keeps the boolean status
-/// of success / failure of the transactions.
-/// Note that the specific details of compute_status are opaque to StateMachineReplication,
-/// which is going to simply pass the results between StateComputer and TxnManager.
-#[derive(Debug, Default, PartialEq, Eq, Clone)]
-pub struct StateComputeResult {
-    pub executed_state: ExecutedState,
-    /// The compute status (success/failure) of the given payload. The specific details are opaque
-    /// for StateMachineReplication, which is merely passing it between StateComputer and
-    /// TxnManager.
-    pub compute_status: Vec<TransactionStatus>,
-}
-
-impl StateComputeResult {
-    pub fn version(&self) -> Version {
-        self.executed_state.version
-    }
-
-    pub fn root_hash(&self) -> HashValue {
-        self.executed_state.state_id
-    }
-
-    pub fn status(&self) -> &Vec<TransactionStatus> {
-        &self.compute_status
-    }
-
-    pub fn has_reconfiguration(&self) -> bool {
-        self.executed_state.validators.is_some()
-    }
-}
-
-/// Executed state derived from StateComputeResult that is maintained with every proposed block.
-/// `state_id`(transaction accumulator root hash) summarized both the information of the version and
-/// the validators.
-#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ExecutedState {
-    /// Tracks the execution state of a proposed block
-    pub state_id: HashValue,
-    /// Version of after executing a proposed block.  This state must be persisted to ensure
-    /// that on restart that the version is calculated correctly
-    pub version: Version,
-    /// If set, this is the validator set that should be changed to if this block is committed.
-    /// TODO [Reconfiguration] the validators are currently ignored, no reconfiguration yet.
-    pub validators: Option<ValidatorSet>,
-}
-
-impl ExecutedState {
-    pub fn state_for_genesis() -> Self {
-        ExecutedState {
-            state_id: *ACCUMULATOR_PLACEHOLDER_HASH,
-            version: 0,
-            validators: None,
-        }
-    }
-}
-
-/// The entire set of data associated with a transaction. In addition to the output generated by VM
-/// which includes the write set and events, this also has the in-memory trees.
-#[derive(Clone, Debug)]
-pub struct TransactionData {
-    /// Each entry in this map represents the new blob value of an account touched by this
-    /// transaction. The blob is obtained by deserializing the previous blob into a BTreeMap,
-    /// applying relevant portion of write set on the map and serializing the updated map into a
-    /// new blob.
-    account_blobs: HashMap<AccountAddress, AccountStateBlob>,
-
-    /// The list of events emitted during this transaction.
-    events: Vec<ContractEvent>,
-
-    /// The execution status set by the VM.
-    status: TransactionStatus,
-
-    /// The in-memory Sparse Merkle Tree after the write set is applied. This is `Rc` because the
-    /// tree has uncommitted state and sometimes `StateVersionView` needs to have a pointer to the
-    /// tree so VM can read it.
-    state_tree: Arc<SparseMerkleTree>,
-
-    /// The in-memory Merkle Accumulator that has all events emitted by this transaction.
-    event_tree: Arc<InMemoryAccumulator<EventAccumulatorHasher>>,
-
-    /// The amount of gas used.
-    gas_used: u64,
-
-    /// The number of newly created accounts.
-    num_account_created: usize,
-
-    /// The transaction info hash if the VM status output was keep, None otherwise
-    txn_info_hash: Option<HashValue>,
-}
-
-impl TransactionData {
-    fn new(
-        account_blobs: HashMap<AccountAddress, AccountStateBlob>,
-        events: Vec<ContractEvent>,
-        status: TransactionStatus,
-        state_tree: Arc<SparseMerkleTree>,
-        event_tree: Arc<InMemoryAccumulator<EventAccumulatorHasher>>,
-        gas_used: u64,
-        num_account_created: usize,
-        txn_info_hash: Option<HashValue>,
-    ) -> Self {
-        TransactionData {
-            account_blobs,
-            events,
-            status,
-            state_tree,
-            event_tree,
-            gas_used,
-            num_account_created,
-            txn_info_hash,
-        }
-    }
-
-    fn account_blobs(&self) -> &HashMap<AccountAddress, AccountStateBlob> {
-        &self.account_blobs
-    }
-
-    fn events(&self) -> &[ContractEvent] {
-        &self.events
-    }
-
-    fn status(&self) -> &TransactionStatus {
-        &self.status
-    }
-
-    fn gas_used(&self) -> u64 {
-        self.gas_used
-    }
-
-    fn num_account_created(&self) -> usize {
-        self.num_account_created
-    }
-
-    fn prune_state_tree(&self) {
-        self.state_tree.prune()
-    }
-
-    pub fn txn_info_hash(&self) -> Option<HashValue> {
-        self.txn_info_hash
-    }
-}
-
-/// Generated by processing VM's output.
-#[derive(Debug, Clone)]
-pub struct ProcessedVMOutput {
-    /// The entire set of data associated with each transaction.
-    transaction_data: Vec<TransactionData>,
-
-    /// The in-memory Merkle Accumulator and state Sparse Merkle Tree after appending all the
-    /// transactions in this set.
-    executed_trees: ExecutedTrees,
-
-    /// If set, this is the validator set that should be changed to if this block is committed.
-    /// TODO [Reconfiguration] the validators are currently ignored, no reconfiguration yet.
-    validators: Option<ValidatorSet>,
-}
-
-impl ProcessedVMOutput {
-    pub fn new(
-        transaction_data: Vec<TransactionData>,
-        executed_trees: ExecutedTrees,
-        validators: Option<ValidatorSet>,
-    ) -> Self {
-        ProcessedVMOutput {
-            transaction_data,
-            executed_trees,
-            validators,
-        }
-    }
-
-    pub fn transaction_data(&self) -> &[TransactionData] {
-        &self.transaction_data
-    }
-
-    pub fn executed_trees(&self) -> &ExecutedTrees {
-        &self.executed_trees
-    }
-
-    pub fn accu_root(&self) -> HashValue {
-        self.executed_trees().txn_accumulator().root_hash()
-    }
-
-    pub fn version(&self) -> Option<Version> {
-        self.executed_trees().version()
-    }
-
-    pub fn validators(&self) -> &Option<ValidatorSet> {
-        &self.validators
-    }
-
-    // This method should only be called by tests.
-    pub fn set_validators(&mut self, validator_set: ValidatorSet) {
-        self.validators = Some(validator_set)
-    }
-
-    pub fn state_compute_result(&self) -> StateComputeResult {
-        let num_leaves = self.executed_trees().txn_accumulator().num_leaves();
-        let version = if num_leaves == 0 { 0 } else { num_leaves - 1 };
-        StateComputeResult {
-            // Now that we have the root hash and execution status we can send the response to
-            // consensus.
-            // TODO: The VM will support a special transaction to set the validators for the
-            // next epoch that is part of a block execution.
-            executed_state: ExecutedState {
-                state_id: self.accu_root(),
-                version,
-                validators: self.validators.clone(),
-            },
-            compute_status: self
-                .transaction_data()
-                .iter()
-                .map(|txn_data| txn_data.status())
-                .cloned()
-                .collect(),
-        }
-    }
-}
 
 /// `Executor` implements all functionalities the execution module needs to provide.
 pub struct Executor<V> {
@@ -689,7 +462,7 @@ where
             return Ok(Some(epoch_change_li));
         }
         ensure!(
-            new_output.validators.is_none(),
+            new_output.validators().is_none(),
             "End of epoch chunk based on local computation but no EoE LedgerInfo provided."
         );
         Ok(None)
@@ -819,15 +592,14 @@ where
             }
         }
 
-        let current_transaction_accumulator = parent_trees
-            .transaction_accumulator
-            .append(&txn_info_hashes);
+        let current_transaction_accumulator =
+            parent_trees.txn_accumulator().append(&txn_info_hashes);
         Ok(ProcessedVMOutput::new(
             txn_data,
-            ExecutedTrees {
-                state_tree: current_state_tree,
-                transaction_accumulator: Arc::new(current_transaction_accumulator),
-            },
+            ExecutedTrees::new_copy(
+                current_state_tree,
+                Arc::new(current_transaction_accumulator),
+            ),
             next_validator_set,
         ))
     }
@@ -918,89 +690,5 @@ where
             .into_iter()
             .filter(|event| *event.key() == reconfig_event_key)
             .collect()
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct ExecutedTrees {
-    /// The in-memory Sparse Merkle Tree representing a specific state after execution. If this
-    /// tree is presenting the latest commited state, it will have a single Subtree node (or
-    /// Empty node) whose hash equals the root hash of the newest Sparse Merkle Tree in
-    /// storage.
-    state_tree: Arc<SparseMerkleTree>,
-
-    /// The in-memory Merkle Accumulator representing a blockchain state consistent with the
-    /// `state_tree`.
-    transaction_accumulator: Arc<InMemoryAccumulator<TransactionAccumulatorHasher>>,
-}
-
-impl From<TreeState> for ExecutedTrees {
-    fn from(tree_state: TreeState) -> Self {
-        ExecutedTrees::new(
-            tree_state.account_state_root_hash,
-            tree_state.ledger_frozen_subtree_hashes,
-            tree_state.version + 1,
-        )
-    }
-}
-
-impl ExecutedTrees {
-    pub fn state_tree(&self) -> &Arc<SparseMerkleTree> {
-        &self.state_tree
-    }
-
-    pub fn txn_accumulator(&self) -> &Arc<InMemoryAccumulator<TransactionAccumulatorHasher>> {
-        &self.transaction_accumulator
-    }
-
-    pub fn version(&self) -> Option<Version> {
-        let num_elements = self.txn_accumulator().num_leaves() as u64;
-        if num_elements > 0 {
-            Some(num_elements - 1)
-        } else {
-            None
-        }
-    }
-
-    pub fn state_id(&self) -> HashValue {
-        self.txn_accumulator().root_hash()
-    }
-
-    pub fn state_root(&self) -> HashValue {
-        self.state_tree().root_hash()
-    }
-
-    pub fn new(
-        state_root_hash: HashValue,
-        frozen_subtrees_in_accumulator: Vec<HashValue>,
-        num_leaves_in_accumulator: u64,
-    ) -> ExecutedTrees {
-        ExecutedTrees {
-            state_tree: Arc::new(SparseMerkleTree::new(state_root_hash)),
-            transaction_accumulator: Arc::new(
-                InMemoryAccumulator::new(frozen_subtrees_in_accumulator, num_leaves_in_accumulator)
-                    .expect("The startup info read from storage should be valid."),
-            ),
-        }
-    }
-
-    pub fn new_empty() -> ExecutedTrees {
-        Self::new(*SPARSE_MERKLE_PLACEHOLDER_HASH, vec![], 0)
-    }
-}
-
-struct ProofReader {
-    account_to_proof: HashMap<HashValue, SparseMerkleProof>,
-}
-
-impl ProofReader {
-    fn new(account_to_proof: HashMap<HashValue, SparseMerkleProof>) -> Self {
-        ProofReader { account_to_proof }
-    }
-}
-
-impl ProofRead for ProofReader {
-    fn get_proof(&self, key: HashValue) -> Option<&SparseMerkleProof> {
-        self.account_to_proof.get(&key)
     }
 }
